Refactor load_eeg_file into shared utility
